Reordenar componentes y refactoring
