--- dream-lab-frontend/src/pages/Dashboard/Dashboard.css
@@ -1,52 +1,0 @@
-.dashboard-main-container {
-    display: grid;
-    gap: max(2vw, 2vh);
-    grid-template-rows: repeat(4, minmax(0, 1fr));
-    height: 100vh;
-    padding: 2vw;
-    padding-top: 13.5vh;
-}
-
-.dashboard-statcards-container {
-    display: grid;
-    gap: max(2vw, 2vh);
-    grid-template-columns: repeat(4, minmax(0, 1fr));
-    grid-row: span 1 / span 1;
-    height: 100%;
-}
-
-.dashboard-graphs-container {
-    display: grid;
-    gap: max(2vw, 2vh);
-    grid-template-rows: repeat(3, minmax(0, 1fr));
-    grid-template-columns: repeat(5, minmax(0, 1fr));
-    grid-row: span 3 / span 3;
-    height: 100%;
-}
-
-.dashboard-grafica-default {
-    background-color: white;
-    border-radius: max(1vw, 2vh);
-    padding: max(1vw, 2vh);
-    overflow-y: auto;
-}
-
-.dashboard-grafica-materiales-container {
-    grid-row: span 2 / span 2;
-    grid-column: span 2 / span 2;
-}
-
-.dashboard-grafica-reservaciones-totales-container {
-    grid-row: span 2 / span 2;
-    grid-column: span 2 / span 2;
-}
-
-.dashboard-grafica-reservaiones-sala-container {
-    grid-row: span 3 / span 3;
-    grid-column: span 1 / span 1;
-}
-
-.dashboard-grafica-disponbilidad-container {
-    grid-row: span 1 / span 1;
-    grid-column: span 4 / span 4;
-}
