+++ dream-lab-frontend/src/pages/Dashboard/components/GraficasDashboard/components/ContenedorDisponibilidadSalas/ContenedorDisponibilidadSalas.jsx
