+++ dream-lab-frontend/src/pages/Dashboard/components/GraficasDashboard/components/ContenedorDisponibilidadSalas/components/EstatusDisponibilidadSala/EstatusDisponibilidadSala.jsx
