+++ dream-lab-frontend/src/pages/Dashboard/components/GraficasDashboard/components/ContenedorGrafica/ContenedorGrafica.jsx
