+++ dream-lab-frontend/src/pages/Dashboard/components/GraficasDashboard/components/GraficaBarras/GraficaBarras.jsx
