+++ dream-lab-frontend/src/pages/Dashboard/components/GraficasDashboard/components/StatCard/StatCard.jsx
